Moved manager files from Scripts folder to Managers subfolder
