--- Assets/_app/Scripts/InputManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: acaee782a742b2643ba369d9ecc8bea7
