+++ Assets/_app/Scripts/Interactables/QuizBook.cs
@@ -1,3 +1,4 @@
+using _app.Scripts.Managers;
 using UnityEngine;
 
 namespace _app.Scripts.Interactables
