+++ Assets/_app/Scripts/Managers/UIManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 597a155d157ff5343aaf5bd657715f24
