+++ Assets/_app/Scripts/Player/PlayerInteract.cs
@@ -1,5 +1,6 @@
 using UnityEngine;
 using _app.Scripts.Interactables;
+using _app.Scripts.Managers;
 
 namespace _app.Scripts.Player
 {
