--- Assets/_app/Scripts/UIManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 663b2621221ae504f935ddc26ccc78b5
